Todolonegroestabancalienteconunpotrón
Probé guardar obstáculos, trayectorias y puertas (todo junto) y lo hace,
pero no guarda las adyacencias todavía. Volví a int.

+++ src/Baldosa.java
@@ -5,7 +5,7 @@ public class Baldosa extends Superficie {
 	private int fila, columna;
 	private boolean pasar = true;
 
-	public Baldosa(double x, double y, double largo, double alto, int fila, int columna) {
+	public Baldosa(int x, int y, int largo, int alto, int fila, int columna) {
 		super(x, y, largo, alto);
 		this.fila = fila;
 		this.columna = columna;
