+++ src/Superficie.java
@@ -1,58 +1,57 @@
-
 import java.awt.geom.Rectangle2D;
 
 public class Superficie {
 
-	protected double x, y, largo, alto;
+	protected int x, y, largo, alto;
 
-	public Superficie(double x, double y, double largo, double alto) {
+	public Superficie(int x, int y, int largo, int alto) {
 		this.x = x;
 		this.y = y;
 		this.largo = largo;
 		this.alto = alto;
 	}
 
-	public boolean contiene(double x, double y) {
+	public boolean contiene(int x, int y) {
 		return x >= this.x && x <= this.x + largo && y >= this.y && y <= this.y + alto;
 	}
 
 	public Rectangle2D getForma() {
-		return new Rectangle2D.Double(x, y, largo, alto);
+		return new Rectangle2D.Float(x, y, largo, alto);
 	}
 
 	public Rectangle2D getInterior() {
-		return new Rectangle2D.Double(x + 1, y + 1, largo - 1, alto - 1);
+		return new Rectangle2D.Float(x + 1, y + 1, largo - 1, alto - 1);
 	}
 
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
-	public double getY() {
+	public int getY() {
 		return y;
 	}
 
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
 
-	public double getLargo() {
+	public int getLargo() {
 		return largo;
 	}
 
-	public void setLargo(double largo) {
+	public void setLargo(int largo) {
 		this.largo = largo;
 	}
 
-	public double getAlto() {
+	public int getAlto() {
 		return alto;
 	}
 
-	public void setAlto(double alto) {
+	public void setAlto(int alto) {
 		this.alto = alto;
 	}
 }
